move AdmissionController construction to separate factory.py
